+++ champ_base/firmware/lib/quadruped_gait/quadruped_gait.cpp
@@ -47,14 +47,14 @@ void QuadrupedGait::transformTrajectory(QuadrupedLeg *leg, float linear_velocity
     float delta_x = transformed_stance.X() - leg->nominal_stance().X();
     float delta_y = transformed_stance.Y() - leg->nominal_stance().Y();
 
-    rotation = atan2(delta_y, delta_x);
-    step_length = sqrt(pow(delta_x, 2) + pow(delta_y, 2));
+    rotation = atan2f(delta_y, delta_x);
+    step_length = sqrtf(pow(delta_x, 2) + pow(delta_y, 2));
 }
 
 void QuadrupedGait::generate(Transformation (&foot_positions)[4], float linear_velocity_x, float linear_velocity_y, float angular_velocity_z)
 {
     float tangential_velocity = angular_velocity_z * (base_->lf->nominal_stance().Y() +  base_->lf->nominal_stance().X());
-    float velocity =  sqrt(pow(linear_velocity_x, 2) + pow(linear_velocity_y + abs(tangential_velocity), 2));
+    float velocity =  sqrtf(pow(linear_velocity_x, 2) + pow(linear_velocity_y + abs(tangential_velocity), 2));
     float step_lengths[4] = {0,0,0,0};
     float trajectory_rotations[4] = {0,0,0,0};    
     float sum_of_steps = 0;
